@@ -31,27 +31,27 @@
     "eventsourcing"
   ],
   "devDependencies": {
-    "@types/jest": "26.0.21",
+    "@types/jest": "27.4.0",
     "@types/knex": "0.16.1",
-    "@types/node": "14.14.35",
-    "@typescript-eslint/eslint-plugin": "4.19.0",
-    "@typescript-eslint/parser": "4.19.0",
-    "eslint": "7.22.0",
-    "jest": "26.6.3",
+    "@types/node": "17.0.8",
+    "@typescript-eslint/eslint-plugin": "5.9.0",
+    "@typescript-eslint/parser": "5.9.0",
+    "eslint": "8.6.0",
+    "jest": "27.4.7",
     "npm-run-all": "4.1.5",
-    "prettier": "2.2.1",
-    "ts-jest": "26.5.4",
-    "ts-node": "9.1.1",
-    "typescript": "4.2.3"
+    "prettier": "2.5.1",
+    "ts-jest": "27.1.2",
+    "ts-node": "10.4.0",
+    "typescript": "4.5.4"
   },
   "dependencies": {
-    "@arpinum/defender": "1.0.3",
-    "@arpinum/promising": "4.2.0",
-    "knex": "0.21.6",
-    "pg": "8.4.1",
-    "pg-query-stream": "3.3.1"
+    "@arpinum/defender": "1.0.4",
+    "@arpinum/promising": "4.2.1",
+    "knex": "0.95.15",
+    "pg": "8.7.1",
+    "pg-query-stream": "4.2.1"
   },
-  "jest":{
+  "jest": {
     "preset": "ts-jest"
   }
 }
